+++ api/src/main/java/com/walking/api/service/path/UpdateRoutePathNameService.java
@@ -1,7 +1,7 @@
-package com.walking.api.web.service.path;
+package com.walking.api.service.path;
 
-import com.walking.api.repository.member.PathFavoritesRepository;
-import com.walking.api.web.service.path.dto.request.PathFavoriteNameRequest;
+import com.walking.api.repository.path.PathFavoritesRepository;
+import com.walking.api.service.dto.request.PathFavoriteNameRequest;
 import com.walking.data.entity.member.MemberEntity;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
